requirements.txt
asttokens==3.0.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
comm==0.2.3
cryptography==47.0.0
curl_cffi==0.13.0
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
frozendict==2.4.7
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
httplib2==0.31.2
idna==3.11
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mdurl==0.1.2
multitasking==0.0.12
nest-asyncio==1.6.0
numpy==2.4.4
oauthlib==3.3.1
packaging==26.1
pandas==2.3.3
parso==0.8.6
peewee==4.0.4
platformdirs==4.4.0
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
pyzmq==27.1.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==15.0.0
six==1.17.0
soupsieve==2.8.3
stack-data==0.6.3
tabulate==0.9.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.6.0
websockets==15.0.1
yfinance==1.2.0
